Move Halo opening scheme to its own file
