+++ src/lib.rs
@@ -33,6 +33,7 @@ mod curve;
 mod fft;
 mod field;
 mod gates;
+pub mod halo;
 mod hash_to_curve;
 mod mds;
 mod partition;
@@ -41,7 +42,7 @@ pub mod plonk_challenger;
 mod plonk_proof;
 mod plonk_recursion;
 pub mod plonk_util;
-mod poly_commit;
+pub mod poly_commit;
 mod polynomial;
 mod pseudorandom;
 mod rescue;
